Consolidate planetary computer data sources and add modality-agnostic base class
